# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Figureheadz collectibles storefront API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  variationId: number;
  productId: number;
  productSlug: string;
  productName: string;
  variationName: string;
  sku: string;
  /** @nullable */
  imageUrl?: string | null;
  unitPriceCents: number;
  quantity: number;
  lineTotalCents: number;
  availableStock: number;
}