# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Figureheadz collectibles storefront API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  variationId: number;
  /** @minimum 1 */
  quantity: number;
}